requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
awscli==1.25.2
botocore==1.27.2
colorama==0.4.4
Deprecated==1.2.13
distlib==0.3.4
Django==4.1.4
djangorestframework==3.11.1
docutils==0.16
factory-boy==3.2.1
Faker==15.3.4
filelock==3.7.1
jmespath==1.0.0
packaging==21.3
pbr==5.9.0
platformdirs==2.5.2
psycopg2==2.9.5
pyasn1==0.4.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==5.4.1
redis==4.3.4
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
stevedore==3.5.0
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wrapt==1.14.1
